Move ASTVisitor and IASTVisitorContext into a separate file
